src/requirements.txt
Alfred-Workflow==1.37.1
algoliasearch==2.0.2
certifi==2019.3.9
chardet==3.0.4
idna==2.8
requests==2.22.0
typing==3.6.6
urllib3==1.25.3
